+++ src/blockly/utils/connection_checker.ts
@@ -1,17 +1,9 @@
 import * as Blockly from "blockly/core";
-import Category from "./category";
-import {Metrics} from "./metrics";
-import {Toolbox} from "./toolbox";
-import {Flyout} from "./flyout";
 
-Blockly.registry.register(Blockly.registry.Type.METRICS_MANAGER, "CustomMetricsManager", Metrics);
-Blockly.registry.register(Blockly.registry.Type.TOOLBOX_ITEM, "category", Category, true);
-
-export default {
-	toolbox: Toolbox,
-	flyoutsVerticalToolbox: Flyout,
-	metricsManager: Metrics,
-	connectionChecker: class extends Blockly.ConnectionChecker {
+Blockly.registry.register(
+    Blockly.registry.Type.CONNECTION_CHECKER,
+    "ScrapConnectionChecker",
+    class extends Blockly.ConnectionChecker {
         doTypeChecks(a: Blockly.Connection, b: Blockly.Connection) {
             const checkArrayOne = a.getCheck();
             const checkArrayTwo = b.getCheck();
@@ -27,4 +19,4 @@ export default {
             );
         }
     }
-};
+);