+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.2.0] - 2022-09-15
+
+### Fixed
+
+- RELEASE_HEADER.md &rarr; RELEASE_HEAD.md
+
+### Added
+
+- Added hex validation to the string field in case of color input.
+- Connection Checker now allows to connect Color and String blocks.
+- About dialog added.
+
+### Removed
+
+- Removed the `hex` block.
+
 ## [2.1.0] - 2022-09-15
 
 ### Changed
@@ -61,8 +77,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reserved words are no longer in Generator's constructor.
 
-[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v2.1.0...HEAD
+[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v2.1.1...HEAD
 [2.1.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.1.0
+[2.1.1]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.1.1
 [2.0.2]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.0.2
 [2.0.1]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.0.1
 [2.0.0]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.0.0